refactor : car 수정
